+++ docs/CHANGELOG.md
@@ -18,12 +18,17 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Add Tint component support for sprites. ([#1756])
 * Remove remaining <N: RealField> type parameter on GameDataBuilder, add Debug derive to LoggerConfig ([#1758])
 * Add `load_from_data_async` to Asset Loader. ([#1753])
+* Add `SerializableFormat` marker trait which is now needed to be implemented for all the formats that are supposed to be serialized. ([#1720])
+
+### Fixed
+* Fix stack overflow on serializing `Box<dyn Format<_>>`. ([#1720])
 
 [#1719]: https://github.com/amethyst/amethyst/pull/1719
 [#1756]: https://github.com/amethyst/amethyst/pull/1756
 [#1733]: https://github.com/amethyst/amethyst/pull/1733
 [#1758]: https://github.com/amethyst/amethyst/pull/1758
 [#1753]: https://github.com/amethyst/amethyst/pull/1753
+[#1720]: https://github.com/amethyst/amethyst/pull/1720
 
 ## [0.11.0] - 2019-06